refactor: Split fields helpers into separate files
