+++ common/helpers/field/index.js
@@ -8,10 +8,10 @@ const {
   set,
 } = require('lodash')
 
-const assessmentCheckboxes = require('../../app/move/fields/common.assessment-checkboxes')
-const explicitAssessmentAnswer = require('../../app/move/fields/common.explicit-assessment-answer')
-const i18n = require('../../config/i18n')
-const componentService = require('../services/component')
+const assessmentCheckboxes = require('../../../app/move/fields/common.assessment-checkboxes')
+const explicitAssessmentAnswer = require('../../../app/move/fields/common.explicit-assessment-answer')
+const i18n = require('../../../config/i18n')
+const componentService = require('../../services/component')
 
 function mapReferenceDataToOption({ id, title, key, conditional, hint }) {
   const option = {
